Refactor masternode into modules
